requirements.txt
aiofiles==24.1.0
aiogram==3.18.0
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
alembic==1.15.1
annotated-types==0.7.0
APScheduler==3.11.0
asyncpg==0.30.0
attrs==25.1.0
black==25.1.0
certifi==2025.1.31
click==8.1.8
dotenv==0.9.9
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
magic-filter==1.0.12
Mako==1.3.9
MarkupSafe==3.0.2
multidict==6.1.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.7
propcache==0.3.0
psycopg2==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.17.0
SQLAlchemy==2.0.38
typing_extensions==4.12.2
tzlocal==5.3.1
yarl==1.18.3
